Add job metadata into manifest
Request parameters, Job ID, server URL.
Most of these won't change after job started, so we save it as a json file at the start of the job.
Go writes the json once, then python reads it and adds into the manifest.
Will be useful for debugging any failed jobs.