Manifest time stamp fix. improved version code.
Timestamp logic to compute the duration of the job into the manifest
was incorrect and threw an exception. This exception got silently caught
causing some info to not be included in the manifest.
Fixed the exception and added log messages.

Version code in refined manifest:
Will work as before when triggered from Github Action,
but when manually building, it will pick branch name and commit hash
instead of just "develop".
Makes it easier to trace results while developing.

# Conflicts:
#	utils/data_utils.py