fix: re-add the global stitch refinement
The old code flow before refactoring was a little bit messy,
and after refactoring it actually wasn't running any global stitch refinement.
Partly since basic_stitch_only needs to be False to reach that code.
(before refactoring we actually ran that even when basic_stitch_only was True)
Also the older global refinement in "run_stitching" is obsolete and
we haven't been using that for very long, so I removed the call to that.

 We would also wanna clean this up to remove run_stitching completely,
but I'd prefer doing that separately from this fix with a bigger cleanup round.