WIP: updated to colmap 3.12. Works in dev container.
Pycolmap 3.12 should support RTX 50xx and also various other improvements. Also has many breaking changes which needed to be changed in our python code too.

I also updated ceres and a bunch of other libs to latest versions.
There's a fix in colmap repo commit 20b2777186654f20745f6c57974924d145bbbd6f
which is absolutely necessary for our pipeline to work. Right now that is not in the latest released 3.12.6 so for now the dockerfile is checking out the exact commit hash.